UI design & fix map teleport

+++ core/src/com/rpsg/rpg/system/ui/ImageButton.java
@@ -122,8 +122,8 @@ public class ImageButton extends Button {
 		updateImage();
 		super.draw(batch, parentAlpha);
 		if(fg!=null){
-			fg.setX(getX()+(getPrefWidth()/2-fg.getWidth()/2));
-			fg.setY(getY()+(getHeight()/2-fg.getHeight()/2));
+			fg.setX((int)(getX()+(getPrefWidth()/2-fg.getWidth()/2)));
+			fg.setY((int)(getY()+(getHeight()/2-fg.getHeight()/2)));
 			fg.setColor(getColor());
 			fg.draw(batch);
 		}
